Updated Graph struct.
The graph struct now exists out of two seperate structs. One for the vertices,
one for the edges. This makes it easier to seperate the parsing of edges from
the parsing of vertices from the gexf file.

When processing the temporal graph, we keep the vertices constant and modify the
edges over time. To prevent storing all edges of all time steps in memory, we
want to parse only the edges for a specific point in time from the gexf file.