Spring embedding now supports changing edges
To support the sliding window, the spring embedding now allows to change the
edges during the spring embedding.

To use this, call the fa2RunOnGraphInStream with the vertices from the graph,
and multiple sets of edges. These edges will be used in sequence for a specified
amount of iterations each.

The implementation uses CUDA streams to perform the spring embedding computation
on a set of edges, while transferring the next set of edges to the device.